Added objparsing functionality

+++ parsers/objparser.go
@@ -11,12 +11,12 @@ import (
 
 // Parse a vertex (normal) string, a list of whitespace-separated
 // floating point numbers.
-func parseVertex(t []string) []float32 {
+func parseVertex(t []string) []float64 {
 	x, _ := strconv.ParseFloat(t[0], 32)
 	y, _ := strconv.ParseFloat(t[1], 32)
 	z, _ := strconv.ParseFloat(t[2], 32)
 
-	return []float32{float32(x), float32(y), float32(z)}
+	return []float64{float64(x), float64(y), float64(z)}
 }
 
 // Parse an element string, a list of whitespace-separated elements.
@@ -24,10 +24,8 @@ func parseVertex(t []string) []float32 {
 // vertex, texture coordinate and normal, respectively.
 func parseElement(t []string) [][3]int32 {
 	e := make([][3]int32, len(t))
-
 	for i := 0; i < len(t); i++ {
 		f := strings.Split(t[i], "/")
-
 		for j := 0; j < len(f); j++ {
 			// for now, just grab the vertex index
 			if x, err := strconv.ParseInt(f[j], 10, 32); err == nil {
@@ -47,20 +45,19 @@ func parseElement(t []string) [][3]int32 {
 }
 
 // ParseObj ...
-func ParseObj(filename string) ([]float32, []float32) {
+func ParseObj(filename string) ([]float64, []float64) {
 	fp, _ := os.Open(filename)
 	scanner := bufio.NewScanner(fp)
 
-	vertices := [][]float32{}
-	normals := [][]float32{}
+	vertices := [][]float64{}
+	normals := [][]float64{}
 	elements := [][3]int32{}
 
-	vertOut := []float32{}
-	normOut := []float32{}
+	vertOut := []float64{}
+	normOut := []float64{}
 
 	for scanner.Scan() {
 		toks := strings.Fields(strings.TrimSpace(scanner.Text()))
-
 		switch toks[0] {
 		case "v":
 			vertices = append(vertices, parseVertex(toks[1:]))
